+++ src/main/java/wooteco/subway/controller/ExceptionController.java
@@ -1,16 +1,16 @@
-package wooteco.subway.controller.line;
+package wooteco.subway.controller;
 
 import java.sql.SQLNonTransientException;
 import org.springframework.dao.DuplicateKeyException;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
-import wooteco.subway.exception.NotFoundLineException;
+import wooteco.subway.exception.line.NotFoundLineException;
+import wooteco.subway.exception.station.NotFoundStationException;
 
 @RestControllerAdvice
-public class LineControllerAdvice {
+public class ExceptionController {
 
     @ExceptionHandler(DuplicateKeyException.class)
     public ResponseEntity<Void> duplicateLineAccessExceptionResponse(final SQLNonTransientException e) {
@@ -18,7 +18,7 @@ public class LineControllerAdvice {
             .build();
     }
 
-    @ExceptionHandler(NotFoundLineException.class)
+    @ExceptionHandler({NotFoundLineException.class, NotFoundStationException.class})
     public ResponseEntity<Void> voidLineDeleteExceptionResponse(final NotFoundLineException e) {
         return ResponseEntity.badRequest()
             .build();
